server/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
textblob>=0.17.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
google-generativeai>=0.3.0
pydantic>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
